+++ packages/data-sql/src/converter.ts
@@ -1,39 +1,5 @@
 import type { AbstractQuery, AbstractQueryFieldNodePrimitive } from '@directus/data/types';
-
-export interface SqlStatementSelectPrimitive {
-	type: 'primitive';
-	table: string;
-	column: string;
-	as?: string;
-}
-
-// export interface SqlStatementSelectFn {
-// 	type: 'fn';
-// 	fn: string;
-// 	args: (string | number | boolean)[];
-// 	table: string;
-// 	column: string;
-// 	as?: string;
-// }
-
-// export interface SqlStatementSelectJson {
-// 	type: 'json';
-// 	table: string;
-// 	column: string;
-// 	as?: string;
-// 	path: string;
-// }
-
-export interface SqlStatement {
-	select: SqlStatementSelectPrimitive[];
-	from: string;
-
-	// limit: {
-	// 	parameterIndex: number;
-	// };
-
-	// parameters: (string | boolean | number)[];
-}
+import type { SqlStatement, SqlStatementSelectPrimitive } from './types.js';
 
 /**
  * @param abstractQuery the abstract query to convert
